Remove any orphaned images after destroy
Handle the following edge case:

An existing deployment has one or more node snapshot images and/or disk
volume images, but no base box image. As a result, the deployment is
registered as "not deployed".

Without this patch, running "sesdev destroy" on that deployment will not
trigger "vagrant destroy" and all the images still existing and
associated with that deployment will be left behind. Then, when the user
tries to create another deployment with the same deployment ID, vagrant
will complain like so:

"Volume for domain is already created. Please run 'vagrant destroy' first."

However, the user is not expected to be able to run "vagrant destroy"
("sesdev destroy" is supposed to do that for him).

Signed-off-by: Nathan Cutler <ncutler@suse.com>